refactor: Add usePersonStore to PersonPage component
The PersonPage component has been updated to use the `usePersonStore` hook from the `person.store.ts` file. This hook provides access to the `firstName` and `lastName` state variables, as well as the `setFirstName` and `setLastName` functions for updating the state. This refactor improves the organization and readability of the PersonPage component by separating the person state management into its own store.